ReadBook

+++ SpMVC_ReadBook/src/main/java/com/biz/read/mapper/BookDao.java
@@ -2,6 +2,6 @@ package com.biz.read.mapper;
 
 import com.biz.read.model.BookVO;
 
-public interface BookDao extends GenericDao<BookVO, Integer> {
+public interface BookDao extends GenericDao<BookVO, Long> {
 
 }
